# EXR: optimize exr image load time