--- a/renderdoc/3rdparty/tinyexr/tinyexr.cpp
+++ b/renderdoc/3rdparty/tinyexr/tinyexr.cpp
@@ -26,5 +26,7 @@
 // compatible functions to be available, which miniz does.
 #define TINYEXR_USE_MINIZ 0
 
+#define TINYEXR_USE_THREAD 1
+
 #define TINYEXR_IMPLEMENTATION
 #include "tinyexr.h"