Faster loading of EXR images:
- Multi-threaded (within one image; tinyexr option)
- Simpler channel planes -> RGBA data swizzling

Tested on two 4K UHD video resolution EXR files, on Ryzen 5950X/VS2022, Release config.
- FP16 file (27MB): 536ms -> 75ms
- FP32 file (69MB): 793ms -> 113ms